fixed build issue
