+++ package.json
@@ -4,14 +4,16 @@
   "private": true,
   "scripts": {
     "dev": "next dev",
-    "build": "next build --no-lint",
+    "build": "npm run clean && next build --no-lint",
     "start": "next start",
     "lint": "next lint",
-    "dev:prod": "cross-env NODE_ENV=production next dev"
+    "dev:prod": "cross-env NODE_ENV=production next dev",
+    "clean": "rimraf .next"
   },
   "dependencies": {
     "@hookform/resolvers": "^5.0.1",
     "clsx": "^2.0.0",
+    "critters": "^0.0.25",
     "date-fns": "^2.30.0",
     "next": "15.3.1",
     "nodemailer": "^6.10.1",
@@ -37,6 +39,7 @@
     "postcss": "^8.4.31",
     "prettier": "^3.1.0",
     "prettier-plugin-tailwindcss": "^0.5.7",
+    "rimraf": "^6.0.1",
     "tailwindcss": "^3.3.5",
     "typescript": "^5"
   }
